+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Throw InvalidArgumentException for non-serializable objects in coercive comparison mode
 * Throw InvalidArgumentException for non-positive count in `Transform::tee()`
 * Throw InvalidArgumentException for step=0 in `Infinite::count()`
+* Throw InvalidArgumentException for negative position in `Reduce::toNth()`
 
 ## v2.0.0 - 2026-03-07